pyproject.toml
[project]
name = "GroupB-tool"
version = "0.1.0"
requires-python = ">=3.13"
dependencies=[
  "pandas>=2.3",
  "gffutils>=0.12",
  "matplotlib>=3.10",
  "seaborn>=0.13",
  "biopython>=1.85"
]
readme = "README.md"
authors = [
  {name = "John T. Hardin", email = "johnhardin2021@gmail.com"},
  {name = "Dom Thompson", email = "dominict1234567@gmail.com"},
  {name = "Hans Henrik Norberg", email = "hhenrikn@outlook.com"},

]
description = "Summarise gene models and output basic GC metrics"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
GroupB-tool="Gene_Model_Summariser.cli:app"

[tool.pytest.ini_options]
testpaths=["tests"]
addopts= ["-q"]

[tool.mypy]
python_version="3.13"
disallow_untyped_defs=true
warn_return_any=true
warn_unused_configs=true
exclude=[
  "tests/",
  "venv/",
]
